moved locations and industries lst in app to dummyData/dropDownItems
