+++ client/src/pages/AddProject.js
@@ -9,6 +9,7 @@ import UploadImages from '../components/UploadImages';
 import FormValidator from '../helpers/form-validation';
 import { withToast } from '../components/Toast';
 import { addProject } from '../api/projects';
+import { locations, industries} from '../dummyData/dropDownItems'
 
 const styles = {
     pageContent: {
@@ -41,23 +42,6 @@ const styles = {
         margin: '20px 0px 10px 0px'
     }
 }
-
-const industries = [
-    { id: 0, name: '' },
-    { id: 1321, name: 'Technology' },
-    { id: 21423, name: 'Marketing' },
-    { id: 42342, name: 'Engineering' },
-    { id: 21342, name: 'Art' },
-    { id: 93082, name: 'Film' }
-]
-
-const locations = [
-    { id: 0, name: '' },
-    { id: 123, name: 'New York' },
-    { id: 431, name: 'California' },
-    { id: 903, name: 'Georgia' },
-    { id: 223, name: 'Florida' }
-]
 
 class AddProject extends Component {
     constructor(props) {
